pyproject.toml
[project]
name = "elgato-downloader"
version = "0.0.1"
description = "CLI tool to download and deploy Elgato Stream Deck Marketplace extensions."
requires-python = ">=3.10"
authors = [{ name = "Lufton" }]
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "urllib3>=1.21.1"
]

[project.scripts]
elgato-downloader = "elgato_downloader:cli"